pyproject.toml
[project]
name = "dc4mem0"
authors = [
    {name="Makarand Bhonsle"}
]
version = "0.1.0"
description = "Salesforce Data Cloud implementation for Mem0 Vector Store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.7.3",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"